feat!: `count!` macro which returns `[_; N]` or `Vec<_>`.
Didn't really liked that we needed 2 different functions when the return type
was the only thing that semantically was different. I guess the only disadvantage
of this, is that you can't no longer specify the generic parameters, but that
should not be an issue; the same `input` and `error` types are used through the
crate's API, so there should not be any inference problems.

refactor: avoid heap allocations when calling `count_const`.

count_const, renamed to `__array_count`, now uses MaybeUninit to allocate a
fixed array, instead of calling `try_into` on a vector.

docs: document all the `utils::nom` items.

BREAKING CHANGE: count_const is now private